+++ phpBB2/search.php
@@ -1141,7 +1141,7 @@ else if ( $search_keywords != '' || $search_author != '' || $search_id )
 
 				for ($k = 0; $k < count($synonym_array); $k++)
 				{ 
-					list($replace_synonym, $match_synonym) = explode(' ', trim(strtolower($synonym_array[$k])));
+					list($replace_synonym, $match_synonym) = explode(' ', trim(phpbb_search_ascii_lower($synonym_array[$k])));
 
 					if ( $replace_synonym == $split_word )
 					{